+++ src/views/CameraPage.vue
@@ -22,6 +22,7 @@
           :showDownload="true"
           :showFullscreen="true"
           :showHistogram="true"
+          :showSolve="true"
           :loading="imageStore.isImageFetching || histogramStore.isProcessing(imageStore.imageData)"
           height="100vh"
           altText="Captured Astrophoto"
@@ -358,7 +359,7 @@ const openModal = (modalType) => {
 // Responsive computed properties
 const iconCenterHere = computed(() => [
   'absolute z-10',
-  !isLandscape.value ? 'top-24 right-40' : 'top-2 right-40', // Kept on right side as it relates to image controls
+  !isLandscape.value ? 'top-24 right-52' : 'top-2 right-52', // Kept on right side as it relates to image controls
 ]);
 
 // Event handlers